requirements.txt
streamlit==1.29.0
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.5
pymongo==4.6.1
python-dotenv==1.0.0
fastapi==0.109.0
uvicorn==0.27.0
requests==2.31.0
openai>=1.10.0
pydantic>=2.5.0
certifi
dnspython
langsmith>=0.0.83,<0.1.0